Adding admin privileges, student access limited

+++ resources/views/students/show.blade.php
@@ -1,7 +1,11 @@
 @extends('layouts.app')
 
 @section('content')
-    <a href="/students">Go Back</a>
+    @if(!Auth::guest())
+      @if(Auth::user()->email == 'admin@gmail.com')
+        <a href="/students">Go Back</a>
+      @endif
+    @endif
     <h1>{{$student->last_name}} {{$student->first_name}}</h1>
     <ol>
       <li>First name: <b>{{$student->first_name}}</b></li>
@@ -10,23 +14,33 @@
       <li>Gender: <b>{{$student->gender}}</b></li>
       <li>Date of birth: <b>{{$student->date_of_birth}}</b></li>
       <li>Place of birth: <b>{{$student->place_of_birth}}</b></li>
-      <li>Personal identification numbre: <b>{{$student->personal_id_number}}</b></li>
+      <li>Personal identification number: <b>{{$student->personal_id_number}}</b></li>
       <li>Email: <b>{{$student->email}}</b></li>
       <li>Phone number: <b>{{$student->phone_number}}</b></li>
       </ol>
-    <a href="/students/{{$student->id}}/edit">Edit student</a>
-    {!!Form::open(['action' => ['StudentsController@destroy', $student->id], 'method' => 'DELETE'])!!}
-      {{Form::submit('Delete student')}}
-    {!!Form::close()!!}
+    @if(!Auth::guest())
+      @if(Auth::user()->email == 'admin@gmail.com')
+        <a href="/students/{{$student->id}}/edit">Edit student</a>
+        {!!Form::open(['action' => ['StudentsController@destroy', $student->id], 'method' => 'DELETE'])!!}
+          {{Form::submit('Delete student')}}
+        {!!Form::close()!!}
+      @endif
+    @endif
     <hr>
     <h2>Subjects</h2>
-    @foreach ($subjects as $subject)
-    <ol>
-      <b>{{$subject->name}}</b>
-      <li>ESPB: {{$subject->espb}}</li>
-      <li>Type: {{$subject->type}}</li>
-      <li>Professor: {{$subject->professor}}</li>
-      <li>Mark: </li>
-    </ol>
-    @endforeach
+    @if(count($subjects) > 0)
+      @foreach ($subjects as $subject)
+      <ol>
+        <b>{{$subject->name}}</b>
+        <li>ESPB: {{$subject->espb}}</li>
+        <li>Type: {{$subject->type}}</li>
+        <li>Professor: {{$subject->professor}}</li>
+        <li>Mark: @foreach($marking as $m)
+          {{$m->mark}}
+        @endforeach</li>
+      </ol>
+      @endforeach
+    @else
+        <p>No subjects found</p>
+    @endif
 @endsection
